starting #428: break up handlers, move clients, start readme
