+++ nbviewer/providers/local/__init__.py
@@ -1,0 +1,1 @@
+from .handlers import LocalFileHandler
